refactoring notifier interface
