+++ operator/controllers/helmrelease_controller_test.go
@@ -22,10 +22,12 @@ import (
 	ctrl "sigs.k8s.io/controller-runtime"
 )
 
-type fakeSlack struct{}
+var sentNotifications []string
 
-func (m *fakeSlack) Send(msg string) error {
-	fmt.Println(msg)
+type fakeNotifier struct{}
+
+func (m *fakeNotifier) Send(msg string) error {
+	sentNotifications = append(sentNotifications, msg)
 	return nil
 }
 
@@ -82,7 +84,7 @@ var _ = Describe("HelmReleaseReconciler", func() {
 	BeforeEach(func() {
 		downloader = new(mockDownloader)
 		helmClient = new(helm.FakeClient)
-		reconciler = NewHelmReleaseReconciler(log, k8sClient, &fakeSlack{}, helmClient, downloader, &recorder, GracePeriod(42), Namespace("test"))
+		reconciler = NewHelmReleaseReconciler(log, k8sClient, &fakeNotifier{}, helmClient, downloader, &recorder, GracePeriod(42), Namespace("test"))
 
 		testRelease = &shipitv1beta1.HelmRelease{
 			TypeMeta: metav1.TypeMeta{
