--- operator/notifications/notifier.go
@@ -1,8 +1,0 @@
-// Notifier
-
-package notifications
-
-// Notifier sends a notification
-type Notifier interface {
-	Send(string) error
-}
